Убрана логика из контроллера, тесты на статр продвижения пользователя
