+++ src/main/java/school/faang/user_service/controller/EventPromotionController.java
@@ -1,0 +1,4 @@
+package school.faang.user_service.controller;
+
+public class EventPromotionController {
+}
