+++ src/main/java/school/faang/user_service/controller/UserPromotionController.java
@@ -1,0 +1,13 @@
+package school.faang.user_service.controller;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/promotion")
+public class UserPromotionController {
+}
