+++ src/main/java/school/faang/user_service/model/promotion/event/EventPromotionPricing.java
@@ -11,11 +11,11 @@ public class EventPromotionPricing {
     private static final Map<Integer, BigDecimal> priorityMultipliers = new HashMap<>();
 
     static {
-        prices.put(EventPromotionType.TEN_PERCENT_OF_USERS.getEventPercentage(), new BigDecimal("30.00"));
-        prices.put(EventPromotionType.TWENTY_FIVE_PERCENT_OF_USERS.getEventPercentage(), new BigDecimal("40.00"));
-        prices.put(EventPromotionType.FORTY_PERCENT_OF_USERS.getEventPercentage(), new BigDecimal("50.00"));
-        prices.put(EventPromotionType.FIFTY_PERCENT_OF_USERS.getEventPercentage(), new BigDecimal("60.00"));
-        prices.put(EventPromotionType.SIXTY_PERCENT_OF_USERS.getEventPercentage(), new BigDecimal("70.00"));
+        prices.put(EventPromotionType.TEN_PERCENT_OF_USERS.getUserPercentage(), new BigDecimal("30.00"));
+        prices.put(EventPromotionType.TWENTY_FIVE_PERCENT_OF_USERS.getUserPercentage(), new BigDecimal("40.00"));
+        prices.put(EventPromotionType.FORTY_PERCENT_OF_USERS.getUserPercentage(), new BigDecimal("50.00"));
+        prices.put(EventPromotionType.FIFTY_PERCENT_OF_USERS.getUserPercentage(), new BigDecimal("60.00"));
+        prices.put(EventPromotionType.SIXTY_PERCENT_OF_USERS.getUserPercentage(), new BigDecimal("70.00"));
 
         priorityMultipliers.put(PromotionPriority.PRIORITY_MINIMAL.getFeedRank(), new BigDecimal("1.0"));
         priorityMultipliers.put(PromotionPriority.PRIORITY_LOW.getFeedRank(), new BigDecimal("2.0"));
