--- src/test/java/school/faang/user_service/service/promotion/EventPromotionTest.java
@@ -1,4 +1,0 @@
-package school.faang.user_service.service.promotion;
-
-public class EventPromotionTest {
-}
