+++ src/test/java/school/faang/user_service/service/promotion/event/EventPromotionTest.java
@@ -1,0 +1,4 @@
+package school.faang.user_service.service.promotion.event;
+
+public class EventPromotionTest {
+}
