+++ src/test/java/school/faang/user_service/service/promotion/user/EndUserPromotionTest.java
@@ -1,0 +1,4 @@
+package school.faang.user_service.service.promotion.user;
+
+public class EndUserPromotionTest {
+}
